requirements.txt
easydict==1.9
numpy==1.17.4
matplotlib==3.1.2
torchvision==0.8.2
torch==1.7.1
Pillow==7.1.2
PyYAML==5.3.1
